Sleep if server is busy
If a delete request returns a status code 429 "Too Many Requests", the script adds the ID back to the list of Tweets to delete and waits one second before trying again.

This should fix #14 . Instead of waiting a second, it would be better to look for a retry-after header. If someone wants to go looking for that, you are welcome.